refactor: update the package structure for workflow-related classes to improve organization and clarity
