+++ src/main/java/org/phong/zenflow/secret/listeners/SecretListener.java
@@ -2,7 +2,7 @@ package org.phong.zenflow.secret.listeners;
 
 import lombok.AllArgsConstructor;
 import org.phong.zenflow.secret.service.SecretLinkSyncService;
-import org.phong.zenflow.workflow.service.event.WorkflowDefinitionUpdatedEvent;
+import org.phong.zenflow.workflow.event.WorkflowDefinitionUpdatedEvent;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.event.TransactionPhase;
