+++ src/main/java/org/phong/zenflow/workflow/dto/WorkflowDefinitionUpdateResult.java
@@ -1,4 +1,4 @@
-package org.phong.zenflow.workflow.service.dto;
+package org.phong.zenflow.workflow.dto;
 
 import org.phong.zenflow.workflow.subdomain.node_definition.definitions.WorkflowDefinition;
 import org.phong.zenflow.workflow.subdomain.schema_validator.dto.ValidationResult;
