+++ src/main/java/org/phong/zenflow/workflow/event/WorkflowDefinitionUpdatedEvent.java
@@ -1,4 +1,4 @@
-package org.phong.zenflow.workflow.service.event;
+package org.phong.zenflow.workflow.event;
 
 import org.phong.zenflow.workflow.subdomain.node_definition.definitions.WorkflowDefinition;
